!fixup Move Address test utils to testutils.rs
